Enable the broker to select a compute node will the least number of open connections

+++ index-common/src/main/java/com/metamx/druid/index/column/ColumnDescriptor.java
@@ -74,9 +74,9 @@ public class ColumnDescriptor
     return parts;
   }
 
-  public int numBytes()
+  public long numBytes()
   {
-    int retVal = 0;
+    long retVal = 0;
 
     for (ColumnPartSerde part : parts) {
       retVal += part.numBytes();
